requirements.txt
streamlit
pypdf
python-docx
transformers
crewai
chromadb
sentence-transformers
langchain
